add test case handle prize no winners

+++ pool/src/state.rs
@@ -115,7 +115,7 @@ pub fn store_lottery_info<S: Storage>(
     bucket(PREFIX_LOTTERY, storage).save(&lottery_id.to_be_bytes(), lottery_info)
 }
 
-pub fn read_lottery_info<S: Storage>(storage: &S, lottery_id: u128) -> Option<LotteryInfo> {
+pub fn read_lottery_info<S: Storage>(storage: &S, lottery_id: u64) -> Option<LotteryInfo> {
     match bucket_read(PREFIX_LOTTERY, storage).load(&lottery_id.to_be_bytes()) {
         Ok(v) => v,
         _ => None,
